adding file watcher
